.github/scripts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
packaging==23.1
    # via
    #   -r requirements.in
    #   build
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pyyaml==6.0.1
    # via -r requirements.in
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
